+++ lib/tabular_a.ex
@@ -5,7 +5,8 @@ defmodule FlowAssertions.TabularA do
   @moduledoc """
   Builders that create functions tailored to tabular tests.
 
-  Here's a example of some tabular tests:
+  Here are some tabular tests for how 
+  `Previously.parse` copes with one of its keyword arguments:
 
       [insert:  :a                  ] |> expect.([een(a: Examples)])
       [insert: [:a, :b      ]       ] |> expect.([een(a: Examples), een(b: Examples)])
@@ -21,13 +22,16 @@ defmodule FlowAssertions.TabularA do
       expect = TabularA.run_and_assert(
         &(Pnode.Previously.parse(&1) |> Pnode.EENable.eens))
 
-  There is also a function that allows assertions about exceptions in a tabular style:
+  There is also a function that allows assertions about exceptions to be written in a tabular style:
 
       raises = TabularA.run_for_exception(&case_clause/2)
       [3, 3] |> raises.(CaseClauseError)
 
-  You can find similar builders in `FlowAssertions.Define.Tabular`. They
-  can be used to test functions that raise `ExUnit.AssertionError` values.
+  ## Related code
+
+  You can find similar builders in
+  `FlowAssertions.Define.Tabular`. They produce functions used to test
+  assertions (like the ones in this package).
 
   ## Beware the typo
 
@@ -48,9 +52,11 @@ defmodule FlowAssertions.TabularA do
 
       expect = TabularA.run_and_assert(&Enum.take/2)
 
-  Quite often, however, it's a function that does some extra work to
-  make the table rows more concise. For example, the following
-  extracts the part of the result that needs to be checked:
+  Quite often, however, it's a function that does some extra work for
+  each table row.
+  For example, the following
+  extracts just the part of the result that needs to be checked. Because of that,
+  the table is less cluttered.
 
       expect = TabularA.run_and_assert(
         &(Pnode.Previously.parse(&1) |> Pnode.EENable.eens))
@@ -60,7 +66,7 @@ defmodule FlowAssertions.TabularA do
 
       :a |> expect.([een(a: Examples)])
 
-  If it had had two or more arguments, those arguments would have had to be passed
+  Functions that take two or more arguments need them to be enclosed
   in a list or tuple:
 
       [:a, :b] |> expect.([een(:a, b: Examples)])
@@ -75,9 +81,9 @@ defmodule FlowAssertions.TabularA do
         &Common.FromPairs.extract_een_values/1,
         &assert_good_enough(&1, in_any_order(&2)))
 
-  In the above case, the second argument is a function that takes both an
-  actual and an expected value. You can instead provide a function that only
-  takes the actual value. In such a case, I typically call the resulting
+  In the above case, the second argument is a function that takes both a
+  computed and an expected value. You can instead provide a function that only
+  takes the computed value. In such a case, I typically call the resulting
   function `pass`:
 
       pass = TabularA.run_and_assert(
@@ -86,8 +92,8 @@ defmodule FlowAssertions.TabularA do
 
       1 |> pass.()
 
-  Beware: a common mistake is to pass in a predicate like `&(&1 == "one passed in")`.
-  Without an assertion, `pass` can never fail.
+  Beware: a common mistake is to use a predicate like `&(&1 == "one passed in")`.
+  Without an assertion, the generated `pass` function can never fail.
   """
   
   def run_and_assert(result_producer, asserter \\ &MiscA.assert_equal/2) do
@@ -126,10 +132,10 @@ defmodule FlowAssertions.TabularA do
 
   Creation looks like this:
 
-        raises = TabularA.run_for_exception(&function_under_test/1)
+      raises = TabularA.run_for_exception(&function_under_test/1)
 
   As with `FlowAssertions.TabularA.run_and_assert/2`, multiple arguments are
-  passed in a list:
+  passed in a list or tuple:
 
       [-1, 3] |> raises.(~r/no negative values/)
 
@@ -143,7 +149,7 @@ defmodule FlowAssertions.TabularA do
 
       [3, 3] |> raises.([CaseClauseError, ~R/no case clause/])
       
-  Note that you can use multiple regular expressions to check different
+  Note that you can put multiple regular expressions in the list to check different
   parts of the message.
 
   The generated function returns the exception, so it can be piped to later